[refactor] LoginRequiredPopover 컴포넌트의 위치를 features 레이어의 auth 도메인으로 이동
- Sidebar에 종속적인 이유로 side-bar 슬라이스에 위치했지만, Navbar에서도 재사용하기 때문에 위치 이동.
